feat(judge): move to `judge/` directory
